+++ es-core/src/resources/TextureResource.cpp
@@ -16,12 +16,8 @@ std::map<TextureResource::TextureKeyType, std::weak_ptr<TextureResource>>
     TextureResource::sTextureMap;
 std::set<TextureResource*> TextureResource::sAllTextures;
 
-TextureResource::TextureResource(const std::string& path,
-                                 bool tile,
-                                 bool dynamic,
-                                 bool linearMagnify,
-                                 bool forceRasterization,
-                                 float scaleDuringLoad)
+TextureResource::TextureResource(
+    const std::string& path, bool tile, bool dynamic, bool linearMagnify, bool forceRasterization)
     : mTextureData(nullptr)
     , mForceLoad(false)
 {
@@ -33,8 +29,6 @@ TextureResource::TextureResource(const std::string& path,
         if (dynamic) {
             data = sTextureDataManager.add(this, tile);
             data->initFromPath(path);
-            if (scaleDuringLoad != 1.0f)
-                data->setScaleDuringLoad(scaleDuringLoad);
             data->setLinearMagnify(linearMagnify);
             data->setForceRasterization(forceRasterization);
             // Force the texture manager to load it using a blocking load.
@@ -44,8 +38,6 @@ TextureResource::TextureResource(const std::string& path,
             mTextureData = std::shared_ptr<TextureData>(new TextureData(tile));
             data = mTextureData;
             data->initFromPath(path);
-            if (scaleDuringLoad != 1.0f)
-                data->setScaleDuringLoad(scaleDuringLoad);
             data->setLinearMagnify(linearMagnify);
             data->setForceRasterization(forceRasterization);
             // Load it so we can read the width/height.
@@ -154,15 +146,14 @@ std::shared_ptr<TextureResource> TextureResource::get(const std::string& path,
                                                       bool forceLoad,
                                                       bool dynamic,
                                                       bool linearMagnify,
-                                                      bool forceRasterization,
-                                                      float scaleDuringLoad)
+                                                      bool forceRasterization)
 {
     std::shared_ptr<ResourceManager>& rm = ResourceManager::getInstance();
 
     const std::string canonicalPath = Utils::FileSystem::getCanonicalPath(path);
     if (canonicalPath.empty()) {
-        std::shared_ptr<TextureResource> tex(new TextureResource(
-            "", tile, false, linearMagnify, forceRasterization, scaleDuringLoad));
+        std::shared_ptr<TextureResource> tex(
+            new TextureResource("", tile, false, linearMagnify, forceRasterization));
         // Make sure we get properly deinitialized even though we do nothing on reinitialization.
         rm->addReloadable(tex);
         return tex;
@@ -178,8 +169,8 @@ std::shared_ptr<TextureResource> TextureResource::get(const std::string& path,
 
     // Need to create it.
     std::shared_ptr<TextureResource> tex;
-    tex = std::shared_ptr<TextureResource>(new TextureResource(
-        key.first, tile, dynamic, linearMagnify, forceRasterization, scaleDuringLoad));
+    tex = std::shared_ptr<TextureResource>(
+        new TextureResource(key.first, tile, dynamic, linearMagnify, forceRasterization));
     std::shared_ptr<TextureData> data = sTextureDataManager.get(tex.get());
 
     // Is it an SVG?